+++ src/prob_conf_mat/experiment_aggregation/interface.py
@@ -8,7 +8,7 @@ from prob_conf_mat.utils import RNG
 def get_experiment_aggregator(
     aggregation: str,
     rng: RNG,
-    **kwargs,  # pyright: ignore[reportUnknownParameterType, reportMissingParameterType]
+    **kwargs,  # pyright: ignore[reportMissingParameterType]
 ) -> ExperimentAggregator:
     """Fetches an `ExperimentAggregator` from its registered name.
 
@@ -23,7 +23,7 @@ def get_experiment_aggregator(
     """
     if aggregation not in AGGREGATION_REGISTRY:
         raise ValueError(
-            (  # noqa: UP034
+            (
                 f"Parameter `aggregation` must be a registered aggregation method. "
                 f"Currently: {aggregation}. Must be one of {set(AGGREGATION_REGISTRY.keys())}"
             )